+++ terraform/modules/happy-ingress-eks/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [2.12.0](https://github.com/chanzuckerberg/happy/compare/happy-ingress-eks-v2.11.0...happy-ingress-eks-v2.12.0) (2024-07-31)
+
+
+### Features
+
+* allow for a path to marked with a fixed-response deny message ([#3455](https://github.com/chanzuckerberg/happy/issues/3455)) ([1cb8dda](https://github.com/chanzuckerberg/happy/commit/1cb8dda981cd09a2354ecba470e397657abb6f0d))
+
 ## [2.11.0](https://github.com/chanzuckerberg/happy/compare/happy-ingress-eks-v2.10.0...happy-ingress-eks-v2.11.0) (2024-01-20)